tiff_metadata.c: parameter safety cleanup
For functions which return constant strings, it is safer to return a
pointer to constant string, rather than have them fill in a buffer,
especially a buffer of unknown size.

The fieldType is an enumerated value; treat it as such. Use an enumerated
lookup table for description and size information.

Signed-off-by: James Perkins <james.perkins@datalogic.com>